+++ src/CacheStorage/PhpFiles.php
@@ -19,7 +19,7 @@ class PhpFiles extends AbstractCache
     {
         $this->configure($configs);
         try {
-            $fileAdapter =  new PhpFilesAdapter('', 0, $this->configs['fs_cache_path']);
+            $fileAdapter = new PhpFilesAdapter('', 0, $this->configs['fs_cache_path']);
             $adapter = !empty($this->configs['use_cache_tags']) ? new TagAwareAdapter($fileAdapter) : $fileAdapter;
             // @codeCoverageIgnoreStart
         } catch (\Exception $e) {