+++ tests/Unit/AbstractRemediation.php
@@ -28,9 +28,11 @@ abstract class AbstractRemediation extends TestCase
         array $methods = []
     ): MockObject {
         switch ($type) {
+            case 'PhpFilesAdapterWithTags':
             case 'PhpFilesAdapter':
                 $class = 'CrowdSec\RemediationEngine\CacheStorage\PhpFiles';
                 break;
+            case 'RedisAdapterWithTags':
             case 'RedisAdapter':
                 $class = 'CrowdSec\RemediationEngine\CacheStorage\Redis';
                 break;