+++ tests/scripts/refresh-decisions-capi.php
@@ -21,19 +21,19 @@ $capiClient = new Watcher($clientConfigs, new FileStorage(__DIR__), null, $logge
 // Init PhpFiles cache storage
 $cacheFileConfigs = [
     'fs_cache_path' => __DIR__ . '/.cache/capi',
-    'use_cache_tags' => false
+    'use_cache_tags' => false,
 ];
 $phpFileCache = new PhpFiles($cacheFileConfigs, $logger);
 // Init Memcached cache storage
 $cacheMemcachedConfigs = [
     'memcached_dsn' => 'memcached://memcached:11211',
-    'use_cache_tags' => false
+    'use_cache_tags' => false,
 ];
 $memcachedCache = new Memcached($cacheMemcachedConfigs, $logger);
 // Init Redis cache storage
 $cacheRedisConfigs = [
     'redis_dsn' => 'redis://redis:6379',
-    'use_cache_tags' => true
+    'use_cache_tags' => true,
 ];
 $redisCache = new Redis($cacheRedisConfigs, $logger);
 // Init CAPI remediation